milestone4 completed

+++ app.py
@@ -1,5 +1,6 @@
 from flask import Flask, jsonify, request
 from flask_pymongo import PyMongo
+from bson import ObjectId
 
 app = Flask(__name__)
 app.config['MONGO_URI'] = 'mongodb+srv://jyotibaisoya:baisoya@cluster0.0gxpf.mongodb.net/ShowZillaDb?retryWrites=true&w=majority'  
@@ -22,6 +23,39 @@ class User:
 
 
 
+class Movie:
+    def __init__(self, name, image, description,duration,price,shows):
+        self.name = name
+        self.image = image
+        self.description = description 
+        self.duration = duration
+        self.price = price
+        self.shows = shows
+
+    def to_dict(self):
+        return {
+            'name': self.name,
+            'image': self.image,
+            'description': self.description,
+            'timing': self.timing,
+            'price': self.price,
+            'shows': self.shows
+        }
+
+class Show:
+    def __init__(self, movie_id, timing, category,price):
+       self.movie_id = movie_id
+       self.timing = timing
+       self.category = category
+       self.price = price 
+
+    def to_dict(self):
+        return {
+            'movie_id': self.movie_id,
+            'timing' : self.timing,
+            'category': self.category,
+            'price' : self.prcie
+        }
 
 
 
@@ -40,6 +74,9 @@ def create_user():
 
     return jsonify({'message': 'User created successfully'}), 201
 
+
+
+
 @app.route("/users/login",methods=["POST"])
 def login_user():
     data = request.get_json()
@@ -55,24 +92,21 @@ def login_user():
     else:
         return jsonify({"message":"user does not exist"})
 
-# @app.route('/users/<user_id>', methods=['PUT'])
-# def update_user(user_id):
-#     user_updates = request.json
-#     user_collection = mongo.db.users
-#     user_collection.update_one({'_id': user_id}, {'$set': user_updates})
-#     return jsonify({'message': f'Dish with ID {user_id} updated successfully'})
 
-@app.route('/users/<string:user_id>', methods=['PUT'])
+
+
+
+@app.route('/users/<user_id>', methods=['PUT'])
 def update_user(user_id):
-    user_updates = request.json
-    user_collection = mongo.db.users
-    result = user_collection.replace_one({'_id': user_id}, user_updates)
+    # Update an existing user in the database based on the request data
+    user_id = ObjectId(user_id)
 
-    if result.modified_count == 1:
-        return jsonify({'message': f'User with ID {user_id} updated successfully'}), 200
-    else:
-        return jsonify({'message': 'User not found'}), 404
-
+    data = request.get_json()
+    user_collection=mongo.db.users
+    updated_user = user_collection.update_one({"_id": user_id}, {"$set": data})
+    if updated_user.modified_count > 0:
+        return jsonify({"message": "User updated successfully"}), 200
+    return jsonify({"message": "User not found"}), 404
 
 
 
@@ -80,15 +114,96 @@ def update_user(user_id):
 
 @app.route('/users/<user_id>', methods=['DELETE'])
 def delete_user(user_id):
+    user_id = ObjectId(user_id)
     # Check if the user exists
-    # user = mongo.db.users.find_one({'_id': user_id})
-    # if not user:
-    #     return jsonify({'message': 'User not found'}), 404
+    user = mongo.db.users.find_one({'_id': user_id})
+    if not user:
+        return jsonify({'message': 'User not found'}), 404
 
     # Delete the user from the database
     mongo.db.users.delete_one({'_id': user_id})
 
     return jsonify({'message': 'User deleted successfully'}), 200
+
+
+# @app.route('/movie/register', methods=['POST'])
+# def create_movie():
+#     data = request.get_json()
+#     name = data.get('name')
+#     image = data.get('image')
+#     timing = data.get('timing')
+#     price = data.get('price')
+
+#     # Create a User object
+#     movie = Movie(name=name, image=image, timing=timing,price=price)
+
+#     # Insert the user document into MongoDB
+#     mongo.db.movies.insert_one(movie.to_dict())
+
+#     return jsonify({'message': 'movie has been added successfully'}), 201
+
+
+# Create a new movie
+@app.route('/movies', methods=['POST'])
+def create_movie():
+    data = request.get_json()
+    movie_collection = mongo.db.movies
+    movie_id = movie_collection.insert_one(data).inserted_id
+    return jsonify({"message": "Movie created successfully", "movie_id": str(movie_id)}), 201
+
+# Get all movies
+@app.route('/movies', methods=['GET'])
+def get_all_movies():
+    movie_collection = mongo.db.movies
+    movies = list(movie_collection.find({}, {"_id": 0}))
+    return jsonify(movies)
+
+# Get a specific movie by movie_id
+@app.route('/movies/<movie_id>', methods=['GET'])
+def get_movie(movie_id):
+    movie_id = ObjectId(movie_id)
+    movie_collection = mongo.db.movies
+    movie = movie_collection.find_one({"_id": movie_id}, {"_id": 0})
+    if movie:
+        return jsonify(movie)
+    return jsonify({"message": "Movie not found"}), 404
+
+# Create a new show for a movie
+# @app.route('/movies/<movie_id>/shows', methods=['POST'])
+# def create_show(movie_id):
+#     data = request.get_json()
+#     movie_id=ObjectId(movie_id)
+#     data["movie_id"] = movie_id
+#     show_collection = mongo.db.shows
+#     show_id = show_collection.insert_one(data).inserted_id
+#     return jsonify({"message": "Show created successfully", "show_id": str(show_id)}), 201
+
+
+@app.route('/movies/<movie_id>/shows', methods=['POST'])
+def create_show(movie_id):
+    data = request.get_json()
+    movie_id=ObjectId(movie_id)
+    data["movie_id"] = movie_id
+    show_collection = mongo.db.shows
+    movie_collection = mongo.db.movies
+    show_id = show_collection.insert_one(data).inserted_id
+
+    # Add the show to the shows array of the respective movie
+    movie_collection.update_one(
+        {"_id": movie_id},
+        {"$push": {"shows": data}}
+    )
+
+    return jsonify({"message": "Show created successfully", "show_id": str(show_id)}), 201
+
+
+# Get all shows for a specific movie
+@app.route('/movies/<movie_id>/shows', methods=['GET'])
+def get_all_shows_for_movie(movie_id):
+    show_collection = mongo.db.shows
+    movie_id = ObjectId(movie_id)
+    shows = list(show_collection.find({"movie_id": movie_id}, {"_id": 0}))
+    return jsonify(shows)
 
 
 if __name__ == '__main__':
